Update 1.24
Updated the magnetize skill to actually magnetize items around you in the field of sight.

I used the code created by FoxMcloud5655 and didn't upgrade it much. The difference from his code is that objects can only be attracted if they are in sight. At the same time, the area of action is average. Thus, I decided not to change the EP costs, since in my case it does not greatly affect the balance. However, the skill looks logically complete.

Thanks to FoxMcloud5655

Co-Authored-By: FoxMcloud5655 <support@foxmcloud.net>